+++ app/src/main/java/com/lelestacia/lelenimexml/core/di/NetworkModule.kt
@@ -1,0 +1,39 @@
+package com.lelestacia.lelenimexml.core.di
+
+import com.lelestacia.lelenimexml.BuildConfig
+import com.lelestacia.lelenimexml.core.network.api.JikanAPI
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NetworkModule {
+
+    @Provides
+    @Singleton
+    fun provideLogger() =
+        if (BuildConfig.DEBUG) {
+            HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.HEADERS)
+        } else HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.NONE)
+
+    @Provides
+    @Singleton
+    fun provideOkHttp(logger: HttpLoggingInterceptor) = OkHttpClient.Builder()
+        .addInterceptor(logger)
+        .build()
+
+    @Provides
+    @Singleton
+    fun provideJikanApi(okHttpClient: OkHttpClient): JikanAPI = Retrofit.Builder()
+        .baseUrl(JikanAPI.BASE_URL)
+        .addConverterFactory(GsonConverterFactory.create())
+        .client(okHttpClient)
+        .build().create(JikanAPI::class.java)
+}
