--- app/src/main/java/com/lelestacia/lelenimexml/core/domain/dto/GenericPagingResponse.kt
@@ -1,12 +1,0 @@
-package com.lelestacia.lelenimexml.core.domain.dto
-
-import com.google.gson.annotations.SerializedName
-
-data class GenericPagingResponse<T>(
-
-    @SerializedName("pagination")
-    val pagination: Pagination,
-
-    @SerializedName("data")
-    val data: T
-)
