--- app/src/main/java/com/lelestacia/lelenimexml/di/NetworkModule.kt
@@ -1,36 +1,0 @@
-package com.lelestacia.lelenimexml.di
-
-import com.lelestacia.lelenimexml.core.domain.remote.JikanAPI
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Provides
-    @Singleton
-    fun provideLogger() = HttpLoggingInterceptor()
-        .setLevel(HttpLoggingInterceptor.Level.HEADERS)
-
-    @Provides
-    @Singleton
-    fun provideOkHttp(logger: HttpLoggingInterceptor) = OkHttpClient.Builder()
-        .addInterceptor(logger)
-        .build()
-
-    @Provides
-    @Singleton
-    fun provideJikanApi(okHttpClient: OkHttpClient): JikanAPI = Retrofit.Builder()
-        .baseUrl(JikanAPI.BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .client(okHttpClient)
-        .build().create(JikanAPI::class.java)
-}
