+++ app/src/main/java/com/lelestacia/lelenimexml/feature_anime/data/model/AnimeEntity.kt
@@ -1,0 +1,4 @@
+package com.lelestacia.lelenimexml.feature_anime.data.model
+
+class AnimeEntity {
+}
