+++ app/src/main/java/com/lelestacia/lelenimexml/feature_anime/domain/model/Anime.kt
@@ -1,0 +1,33 @@
+package com.lelestacia.lelenimexml.feature_anime.domain.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class Anime(
+    val malId: Int,
+    val images: String,
+    val trailer: Trailer?,
+    val title: String,
+    val titleEnglish: String?,
+    val titleJapanese: String?,
+    val type: String,
+    val episodes: Int?,
+    val status: String,
+    val rating: String,
+    val score: Double?,
+    val scoredBy: Int?,
+    val rank: Int,
+    val synopsis: String?,
+    val season: String?,
+    val year: Int,
+    val genres: List<String>,
+) : Parcelable {
+    @Parcelize
+    data class Trailer(
+        val youtubeId: String?,
+        val url: String?,
+        val images: String?
+    ) : Parcelable
+}
