+++ app/src/main/java/com/lelestacia/lelenimexml/feature_anime/domain/repository/AnimeRepository.kt
@@ -3,16 +3,14 @@ package com.lelestacia.lelenimexml.feature_anime.domain.repository
 import androidx.paging.PagingData
 import com.lelestacia.lelenimexml.core.domain.dto.animefull.AnimeFUll
 import com.lelestacia.lelenimexml.core.utililty.NetworkResponse
-import com.lelestacia.lelenimexml.feature_anime.domain.model.AnimeCard
+import com.lelestacia.lelenimexml.feature_anime.domain.model.Anime
 import kotlinx.coroutines.flow.Flow
 
 interface AnimeRepository {
 
     suspend fun getAnimeById(animeID: Int): NetworkResponse<AnimeFUll>
 
-    fun seasonAnimePagingData(): Flow<PagingData<AnimeCard>>
+    fun seasonAnimePagingData(): Flow<PagingData<Anime>>
 
-    fun searchAnimeByTitle(): Flow<PagingData<AnimeCard>>
-
-    fun searchNewQuery(newQuery: String)
+    fun searchAnimeByTitle(query: String): Flow<PagingData<Anime>>
 }
