--- app/src/main/java/com/lelestacia/lelenimexml/feature_anime/ui/layout/MyListAnimeFragment.kt
@@ -1,19 +1,0 @@
-package com.lelestacia.lelenimexml.feature_anime.ui.layout
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.lelestacia.lelenimexml.R
-
-class MyListAnime : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_my_list_anime, container, false)
-    }
-}
