+++ app/src/main/java/com/lelestacia/lelenimexml/feature_anime/ui/view/gallery/GalleryFragment.kt
@@ -1,0 +1,42 @@
+package com.lelestacia.lelenimexml.feature_anime.ui.view.gallery
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.lelestacia.lelenimexml.databinding.FragmentGalleryBinding
+
+class GalleryFragment : Fragment() {
+
+    private var _binding: FragmentGalleryBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val galleryViewModel =
+            ViewModelProvider(this).get(GalleryViewModel::class.java)
+
+        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
+        val root: View = binding.root
+
+        val textView: TextView = binding.textGallery
+        galleryViewModel.text.observe(viewLifecycleOwner) {
+            textView.text = it
+        }
+        return root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
