+++ app/src/main/java/com/lelestacia/lelenimexml/feature_anime/ui/view/gallery/GalleryViewModel.kt
@@ -1,0 +1,13 @@
+package com.lelestacia.lelenimexml.feature_anime.ui.view.gallery
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class GalleryViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is gallery Fragment"
+    }
+    val text: LiveData<String> = _text
+}
