+++ app/src/main/java/com/lelestacia/lelenimexml/feature_anime/ui/view/slideshow/SlideshowViewModel.kt
@@ -1,0 +1,13 @@
+package com.lelestacia.lelenimexml.feature_anime.ui.view.slideshow
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SlideshowViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is slideshow Fragment"
+    }
+    val text: LiveData<String> = _text
+}
