+++ app/src/main/res/layout/fragment_my_list_anime.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature_anime.ui.layout.MyListAnime">
+    tools:context=".feature_anime.ui.view.MyListAnime">
 
     <TextView
         android:layout_width="match_parent"
