+++ app/src/main/res/layout/fragment_season_anime.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/background"
-    tools:context=".feature_anime.ui.layout.SeasonAnimeFragment">
+    tools:context=".feature_anime.ui.view.SeasonAnimeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_season_anime"
